Update use case test init
